// src/models/signup-request.models.ts
export class SignUpRequest {
    name: string;
    email: string;
    password: string;
    mobile_number: string;
    image_url: string;
    role: string;

    constructor(name: string, email: string, password: string, mobile_number: string) {
        this.name = name;
        this.email = email;
        this.password = password;
        this.mobile_number = mobile_number;
        this.role = "customer";
    }
}